+++ decomp.c
@@ -153,14 +153,14 @@ int decomp( void )
 				// now move what's left to beginning of next OUTBLOCK
 				if( Debug >= 2 )
 					fprintf( stderr,
-					  "DC: outData = %p, outPtr = %p, outPtr-outEnd = %ld\n",
-					  outData, outPtr, outEnd - outPtr );
+					  "DC: outData = %p, outPtr = %p, outPtr-outEnd = %d\n",
+					  outData, outPtr, (int) (outEnd - outPtr) );
 				memmove( outData, outPtr, (outEnd - outPtr) );
 				outPtr -= OUTBLOCK;
 				if( Debug >= 2 )	
 					fprintf( stderr,
-					  "DC: outData = %p, outPtr = %p, outPtr-outEnd = %ld\n",
-					  outData, outPtr, outEnd - outPtr );
+					  "DC: outData = %p, outPtr = %p, outPtr-outEnd = %d\n",
+					  outData, outPtr, (int) (outEnd - outPtr) );
 			}
 		} while( available >= (COMPBLOCK +1 ));
 
@@ -179,8 +179,8 @@ int decomp( void )
 	}
 	else if( lenWritten != (outPtr - outData) )
 	{
-		fprintf( stderr, "Short write!  Wrote %d, had %ld\n", lenWritten,
-		  outPtr - outData );
+		fprintf( stderr, "Short write!  Wrote %d, had %d\n", lenWritten,
+		  (int) (outPtr - outData) );
 		return 7;
 	}
